pyproject.toml
[tool.poetry]
name = "project_name"
version = "0.1.0"
description = "project_description"
authors = ["author_name"]

[tool.poetry.dependencies]
python = "3.9.*"
pandas = "^1"
xlrd = "^2"
scikit-learn = "^1"
spacy = {extras = ["cuda117", "transformers", "lookups"], version = "^3"}

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "*"
isort = "*"
nbstripout = "*"
pydocstyle = {extras = ["toml"], version = "*"}
notebook = "*"
pre-commit = "*"
rich = "*"
matplotlib = "*"
seaborn = "*"
streamlit = "*"
spacy-streamlit = "*"
dash = "*"
dash-bootstrap-components = "*"
azure-identity = "*"
annoy = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = 'black'

[tool.pydocstyle]
convention = "google"
add-ignore = "D100,D101,D102,D103,D104,D105,D106,D107"
